Move about screen to :ui:preferences
