+++ ui/preferences/src/main/res/layout/settings_activity.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.fragment.app.FragmentContainerView xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.fragment.app.FragmentContainerView
+    xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/settingsContainer"
     android:layout_width="match_parent"
     android:layout_height="match_parent" />
